fix(echo): break Annotation import cycles for check-cycles
Split dictionary and entity get/set into dedicated modules, add an
isEntity guard, and replace barrel imports with direct module paths.
