+++ packages/core/echo/echo/src/internal/Annotation/annotations.ts
@@ -14,13 +14,10 @@ import { type Primitive } from '@dxos/util';
 
 import { type Mutable } from '../common/proxy';
 import { type AnyProperties, EntityKind, TypeId, getSchema } from '../common/types';
+import { Key } from './dictionary';
 import { type AnnotationHelper, createAnnotationHelper } from './util';
 import type * as Annotation from '../../Annotation';
-import * as entityInternal from '../Entity';
 import { todo } from '@dxos/debug';
-import type * as Entity from '../../Entity';
-import type { Types } from 'effect';
-import { getMetaChecked } from '../common';
 
 const ANNOTATION_TYPE_ID: Annotation.TypeId = '~@dxos/echo/Annotation' as const;
 
@@ -595,49 +592,11 @@ export const setDescription = (entity: Mutable<AnyProperties>, description: stri
 };
 
 
-export const Key = Schema.String.pipe(Schema.brand('~@dxos/echo/AnnotationKey'));
-
-export const Dictionary = Schema.Record({ key: Key, value: Schema.Unknown });
-
-export const get = <T>(target: Entity.Unknown, annotation: Annotation.Annotation<T>): Option.Option<T> => {
-  if (entityInternal.isEntity(target)) {
-    const meta = getMetaChecked(target);
-    if (!meta.annotations) {
-      return Option.none();
-    }
-    return getDictionary(meta.annotations, annotation);
-  } else {
-    throw new TypeError('Target is not an annotation target.');
-  }
-};
-
-export const set = <T>(target: Entity.Unknown, annotation: Annotation.Annotation<T>, value: T): void => {
-  if (entityInternal.isEntity(target)) {
-    const meta = getMetaChecked(target);
-    if (!meta.annotations) {
-      meta.annotations = {};
-    }
-    setDictionary(meta.annotations, annotation, value);
-  } else {
-    throw new TypeError('Target is not an annotation target.');
-  }
-};
+export { Dictionary, Key, getDictionary, setDictionary } from './dictionary';
 
 export const getFromAst = <T>(ast: SchemaAST.AST, annotation: Annotation.Annotation<T>): Option.Option<T> => {
   return SchemaAST.getAnnotation<PropertyMetaAnnotation>(PropertyMetaAnnotationId)(ast).pipe(
     Option.flatMap((meta) => Option.fromNullable(meta[annotation.key])),
     Option.map(Schema.decodeUnknownSync(annotation.schema)),
   );
-};
-
-export const getDictionary = <T>(values: Annotation.Dictionary, annotation: Annotation.Annotation<T>): Option.Option<T> => {
-  if (!(annotation.key in values)) {
-    return Option.none();
-  }
-
-  return Function.pipe(values[annotation.key], Schema.decodeUnknownSync(annotation.schema), Option.some);
-};
-
-export const setDictionary = <T>(values: Types.Mutable<Annotation.Dictionary>, annotation: Annotation.Annotation<T>, value: T): void => {
-  values[annotation.key] = Schema.encodeSync(annotation.schema)(value);
 };
