+++ packages/core/echo/echo/src/internal/Annotation/index.ts
@@ -2,6 +2,8 @@
 // Copyright 2024 DXOS.org
 //
 
-export * from './annotations';
-export * from './sorting';
+export * from './dictionary';
 export * from './util';
+export * from './sorting';
+export * from './annotations';
+export * from './entity-dictionary';
