+++ packages/core/echo/echo/src/internal/Entity/guard.ts
@@ -1,0 +1,16 @@
+//
+// Copyright 2026 DXOS.org
+//
+
+import { KindId } from '../common/types';
+import type * as Entity from '../../Entity';
+
+/**
+ * Returns true if the value is an ECHO entity instance (object or relation).
+ */
+export const isEntity = (value: unknown): value is Entity.Unknown => {
+  if (typeof value !== 'object' || value === null) {
+    return false;
+  }
+  return (value as any)[KindId] !== undefined;
+};
