+++ packages/core/echo/echo/src/internal/Entity/type-uri.ts
@@ -8,7 +8,7 @@ import { raise } from '@dxos/debug';
 import { assertArgument } from '@dxos/invariant';
 import { DXN, URI } from '@dxos/keys';
 
-import { getSchemaURI, getTypename } from '../Annotation';
+import { getSchemaURI, getTypename } from '../Annotation/annotations';
 import { type AnyEntity, InstancePhantomId, KindId, TypeId, getStaticTypeSchema } from '../common/types';
 import { getUri as getUriFromEntity } from './api';
