+++ packages/core/echo/echo/src/internal/common/types/meta.ts
@@ -10,7 +10,7 @@ import { type Comparator, intersection } from '@dxos/util';
 
 import type * as Entity from '../../../Entity';
 import { type AnyProperties } from './base';
-import * as internalAnnotation from '../../Annotation';
+import { Dictionary } from '../../Annotation/dictionary';
 
 /**
  * Property name for meta when object is serialized to JSON.
@@ -55,7 +55,7 @@ export const EntityMetaSchema = Schema.Struct({
    * Dictionary of annotations to this entity.
    */
   // TODO(dmaretskyi): Make required.
-  annotations: Schema.optional(internalAnnotation.Dictionary),
+  annotations: Schema.optional(Dictionary),
 });
 
 export type EntityMeta = Schema.Schema.Type<typeof EntityMetaSchema>;
